+++ packages/contracts/package.json
@@ -12,6 +12,7 @@
     "typecheck": "tsc --noEmit",
     "node": "hardhat node",
     "deploy:setup": "NODE_OPTIONS='--experimental-loader ts-node/esm/transpile-only --no-warnings=ExperimentalWarning' hardhat run scripts/deploy-setup.ts --network local",
+    "deploy:setup:l2": "NODE_OPTIONS='--experimental-loader ts-node/esm/transpile-only --no-warnings=ExperimentalWarning' hardhat run scripts/deploy-setup.ts --network local",
     "deploy:setup:remote": "NODE_OPTIONS='--experimental-loader ts-node/esm/transpile-only --no-warnings=ExperimentalWarning' hardhat run scripts/deploy-setup.ts --network remote",
     "deploy:upgrade": "NODE_OPTIONS='--experimental-loader ts-node/esm/transpile-only --no-warnings=ExperimentalWarning' hardhat run scripts/start-upgrade.ts $HANDLER --network local",
     "clean": "hardhat clean",
@@ -20,12 +21,15 @@
     "test:local": "NODE_OPTIONS='--experimental-loader ts-node/esm/transpile-only --no-warnings=ExperimentalWarning' hardhat test --network local"
   },
   "devDependencies": {
+    "@matterlabs/hardhat-zksync-deploy": "^1.5.0",
+    "@matterlabs/hardhat-zksync-node": "^1.1.1",
+    "@matterlabs/hardhat-zksync-solc": "^1.2.4",
     "@nomicfoundation/hardhat-viem": "2.0.3",
     "@openzeppelin-v4/contracts": "npm:@openzeppelin/contracts@4",
     "@openzeppelin-v4/contracts-upgradeable": "npm:@openzeppelin/contracts-upgradeable@4",
     "@openzeppelin/contracts": "^5.0.2",
     "@openzeppelin/contracts-upgradeable": "^5.0.2",
-    "@repo/typescript-config": "workspace:*",
+    "@repo/typescript-config": "workspace:^",
     "@types/chai": "^4.3.19",
     "@types/node": "^22.5.1",
     "chai": "^5.1.1",
@@ -36,6 +40,7 @@
     "ora": "^8.0.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.5.4",
-    "viem": "2.17.5"
+    "viem": "2.17.5",
+    "zksync-ethers": "^6.12.0"
   }
 }
